+++ R/load_constructors.R
@@ -7,14 +7,36 @@
 #' @export
 #' @return A tibble with one row per constructor
 load_constructors <- function() {
-  url <- "constructors.json?limit=300"
-  data <- get_ergast_content(url)
+  lim <- 100
+  url <- glue::glue("constructors.json?limit={lim}", lim = lim)
+  data <- get_jolpica_content(url)
 
   if (is.null(data)) {
     return(NULL)
   }
 
-  return(data$MRData$ConstructorTable$Constructors %>%
+  total <- data$MRData$total %>% as.numeric()
+  offset <- data$MRData$offset %>% as.numeric()
+
+  full <- data$MRData$ConstructorTable$Constructors
+
+  # Iterate over the request until completed
+  while (offset + lim <= total) {
+    offset <- offset + lim
+
+    url <- glue::glue("constructors.json?limit={lim}&offset={offset}",
+      lim = lim, offset = offset
+    )
+    data <- get_jolpica_content(url)
+
+    if (is.null(data)) {
+      return(NULL)
+    }
+
+    full <- dplyr::bind_rows(full, data$MRData$ConstructorTable$Constructors)
+  }
+
+  return(full %>%
     dplyr::select("constructorId", "name", "nationality") %>%
     janitor::clean_names())
 }