+++ R/load_sprint.R
@@ -22,7 +22,7 @@ load_sprint <- function(season = get_current_season(), round = "last") {
     season = season, round = round
   )
 
-  data <- get_ergast_content(url)
+  data <- get_jolpica_content(url)
 
   if (is.null(data)) {
     return(NULL)