+++ R/load_standings.R
@@ -25,7 +25,7 @@ load_standings <- function(season = get_current_season(), round = "last", type =
     season = season, round = round, type = type
   )
 
-  data <- get_ergast_content(url)
+  data <- get_jolpica_content(url)
 
   if (is.null(data)) {
     return(NULL)