+++ tests/testthat/test-load_constructors.R
@@ -7,7 +7,7 @@ test_that("load_constructors works", {
   dir.create(file.path(tempdir(), "tst_load_constructors"), recursive = TRUE)
   withr::local_options(f1dataR.cache = file.path(tempdir(), "tst_load_constructors"))
 
-  skip_if_no_ergast()
+  skip_if_no_jolpica()
 
   constructors <- load_constructors()
 
@@ -15,6 +15,7 @@ test_that("load_constructors works", {
 
   expect_equal(ncol(constructors), 3)
   expect_equal(constructors[1, ]$constructor_id, "adams")
+  expect_true(nrow(unique(constructors)) >= 212)
 })
 
 test_that("load_constructors works without internet", {
@@ -34,7 +35,7 @@ test_that("load_constructors works without internet", {
     suppressWarnings({
       suppressMessages({
         httptest2::without_internet({
-          expect_message(load_constructors(), "f1dataR: Error getting data from Ergast")
+          expect_message(load_constructors(), "f1dataR: Error getting data from Jolpica")
           expect_null(load_constructors())
         })
       })