+++ tests/testthat/test-load_drivers.R
@@ -6,7 +6,7 @@ test_that("Drivers Load works", {
   dir.create(file.path(getwd(), "tst_load_drivers"), recursive = TRUE)
   withr::local_options(f1dataR.cache = file.path(getwd(), "tst_load_drivers"))
 
-  skip_if_no_ergast()
+  skip_if_no_jolpica()
 
   drivers_2021 <- load_drivers(2021)
 
@@ -40,7 +40,7 @@ test_that("load_drivers works without internet", {
     suppressWarnings({
       suppressMessages({
         httptest2::without_internet({
-          expect_message(load_drivers(2021), "f1dataR: Error getting data from Ergast")
+          expect_message(load_drivers(2021), "f1dataR: Error getting data from Jolpica")
           expect_null(load_drivers(2021))
         })
       })