+++ tests/testthat/test-load_quali.R
@@ -7,7 +7,7 @@ test_that("load_quali works", {
   dir.create(file.path(tempdir(), "tst_load_quali"), recursive = TRUE)
   withr::local_options(f1dataR.cache = file.path(tempdir(), "tst_load_quali"))
 
-  skip_if_no_ergast()
+  skip_if_no_jolpica()
 
   quali_2021_1 <- load_quali(2021, 1)
 
@@ -44,7 +44,7 @@ test_that("load_quali works without internet", {
     suppressWarnings({
       suppressMessages({
         httptest2::without_internet({
-          expect_message(load_quali(2021, 1), "f1dataR: Error getting data from Ergast")
+          expect_message(load_quali(2021, 1), "f1dataR: Error getting data from Jolpica")
           expect_null(load_quali(2021, 1))
         })
       })