+++ tests/testthat/test-load_schedule.R
@@ -7,7 +7,7 @@ test_that("load_schedule works", {
   dir.create(file.path(tempdir(), "tst_load_schedule"), recursive = TRUE)
   withr::local_options(f1dataR.cache = file.path(tempdir(), "tst_load_schedule"))
 
-  skip_if_no_ergast()
+  skip_if_no_jolpica()
 
   schedule_2021 <- load_schedule(2021)
 
@@ -46,7 +46,7 @@ test_that("load_schedule works without internet", {
     suppressWarnings({
       suppressMessages({
         httptest2::without_internet({
-          expect_message(load_schedule(2021), "f1dataR: Error getting data from Ergast")
+          expect_message(load_schedule(2021), "f1dataR: Error getting data from Jolpica")
           expect_null(load_schedule(2021))
         })
       })