+++ tests/testthat/test-load_standings.R
@@ -6,7 +6,7 @@ test_that("load_standings works", {
   dir.create(file.path(tempdir(), "tst_load_standings"), recursive = TRUE)
   withr::local_options(f1dataR.cache = file.path(tempdir(), "tst_load_standings"))
 
-  skip_if_no_ergast()
+  skip_if_no_jolpica()
 
   standings_2021 <- load_standings(2021)
 
@@ -38,7 +38,7 @@ test_that("load_standings works without internet", {
     suppressWarnings({
       suppressMessages({
         httptest2::without_internet({
-          expect_message(load_standings(2021), "f1dataR: Error getting data from Ergast")
+          expect_message(load_standings(2021), "f1dataR: Error getting data from Jolpica")
           expect_null(load_standings(2021))
         })
       })